Move more code around
