+++ crates/egui/src/text_selection/text_cursor_state.rs
@@ -1,11 +1,72 @@
 //! Text cursor changes/interaction, without modifying the text.
 
 use epaint::text::{cursor::*, Galley};
-use text_edit::state::TextCursorState;
 
 use crate::*;
 
 use super::{CCursorRange, CursorRange};
+
+/// The state of a text cursor selection.
+///
+/// Used for [`crate::TextEdit`] and [`crate::Label`].
+#[derive(Clone, Copy, Debug, Default)]
+#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
+#[cfg_attr(feature = "serde", serde(default))]
+pub struct TextCursorState {
+    cursor_range: Option<CursorRange>,
+
+    /// This is what is easiest to work with when editing text,
+    /// so users are more likely to read/write this.
+    ccursor_range: Option<CCursorRange>,
+}
+
+impl TextCursorState {
+    pub fn is_empty(&self) -> bool {
+        self.cursor_range.is_none() && self.ccursor_range.is_none()
+    }
+
+    /// The the currently selected range of characters.
+    pub fn char_range(&self) -> Option<CCursorRange> {
+        self.ccursor_range.or_else(|| {
+            self.cursor_range
+                .map(|cursor_range| cursor_range.as_ccursor_range())
+        })
+    }
+
+    pub fn range(&mut self, galley: &Galley) -> Option<CursorRange> {
+        self.cursor_range
+            .map(|cursor_range| {
+                // We only use the PCursor (paragraph number, and character offset within that paragraph).
+                // This is so that if we resize the [`TextEdit`] region, and text wrapping changes,
+                // we keep the same byte character offset from the beginning of the text,
+                // even though the number of rows changes
+                // (each paragraph can be several rows, due to word wrapping).
+                // The column (character offset) should be able to extend beyond the last word so that we can
+                // go down and still end up on the same column when we return.
+                CursorRange {
+                    primary: galley.from_pcursor(cursor_range.primary.pcursor),
+                    secondary: galley.from_pcursor(cursor_range.secondary.pcursor),
+                }
+            })
+            .or_else(|| {
+                self.ccursor_range.map(|ccursor_range| CursorRange {
+                    primary: galley.from_ccursor(ccursor_range.primary),
+                    secondary: galley.from_ccursor(ccursor_range.secondary),
+                })
+            })
+    }
+
+    /// Sets the currently selected range of characters.
+    pub fn set_char_range(&mut self, ccursor_range: Option<CCursorRange>) {
+        self.cursor_range = None;
+        self.ccursor_range = ccursor_range;
+    }
+
+    pub fn set_range(&mut self, cursor_range: Option<CursorRange>) {
+        self.cursor_range = cursor_range;
+        self.ccursor_range = None;
+    }
+}
 
 impl TextCursorState {
     /// Handle clicking and/or dragging text.
