+++ crates/egui/src/widgets/text_edit/mod.rs
@@ -1,14 +1,9 @@
 mod builder;
-pub mod cursor_interaction;
-mod cursor_range;
 mod output;
 mod state;
 mod text_buffer;
 
-#[cfg(feature = "accesskit")]
-pub mod accesskit_text;
-
 pub use {
-    builder::TextEdit, cursor_range::*, output::TextEditOutput, state::TextCursorState,
+    crate::text_selection::TextCursorState, builder::TextEdit, output::TextEditOutput,
     state::TextEditState, text_buffer::TextBuffer,
 };
